requirements.txt
annotated-types==0.6.0
pyaml==23.9.7
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
typing_extensions==4.8.0
